// src/types/midtrans.ts
export interface MidtransTransactionRequest {
  orderId: string;
  grossAmount: number;
  customerDetails: {
    first_name: string;
    last_name?: string;
    email: string;
    phone: string;
  };
  itemDetails: Array<{
    id: string;
    price: number;
    quantity: number;
    name: string;
    category?: string;
  }>;
}

export interface MidtransTransactionResponse {
  token: string;
  redirect_url: string;
}

export interface MidtransNotification {
  transaction_time: string;
  transaction_status: string;
  transaction_id: string;
  status_message: string;
  status_code: string;
  signature_key: string;
  settlement_time?: string;
  payment_type: string;
  order_id: string;
  merchant_id: string;
  gross_amount: string;
  fraud_status: string;
  currency: string;
}